[refactor♻️] PostModifyCommand dto 패키지로 변경(#168)
