+++ CultureSpot-Domain/src/main/java/com/culturespot/culturespotdomain/core/post/service/PostService.java
@@ -1,7 +1,7 @@
 package com.culturespot.culturespotdomain.core.post.service;
 
 import com.culturespot.culturespotdomain.core.post.entity.Post;
-import com.culturespot.culturespotdomain.core.post.entity.PostModifyCommand;
+import com.culturespot.culturespotdomain.core.post.dto.PostModifyCommand;
 import com.culturespot.culturespotdomain.core.user.entity.User;
 import org.springframework.transaction.annotation.Transactional;
